+++ src/main/java/net/silentchaos512/scalinghealth/ScalingHealth.java
@@ -18,7 +18,7 @@ import java.util.Random;
 public class ScalingHealth {
     public static final String MOD_ID = "scalinghealth";
     public static final String MOD_NAME = "Scaling Health";
-    public static final String VERSION = "2.4.1";
+    public static final String VERSION = "2.4.2";
     public static final String RESOURCE_PREFIX = MOD_ID + ":";
 
     public static final Random random = new Random();