+++ src/main/java/net/silentchaos512/scalinghealth/capability/IPetData.java
@@ -1,0 +1,11 @@
+package net.silentchaos512.scalinghealth.capability;
+
+import net.minecraft.entity.passive.TameableEntity;
+
+public interface IPetData {
+    void addHealth(double hp, TameableEntity pet );
+
+    float getBonusHealth();
+
+    void tick(TameableEntity pet );
+}