+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.4.2] - 2020-03-05
+### Added
+- Pets will now accept heart crystals and it increases their hp. The amount is configurable.
+### Fixed
+- Being afk at max difficulty no longer displays a message (or does anything)
+
 ## [2.4.1] - 2020-02-25
 ### Added
 - Key bind option to display the difficulty bar